# feat(pgwire): replace BoxStream in PgResult with static type