--- a/src/frontend/src/handler/drop_sink.rs
+++ b/src/frontend/src/handler/drop_sink.rs
@@ -18,6 +18,7 @@ use risingwave_common::error::Result;
 use risingwave_sqlparser::ast::ObjectName;
 
 use super::privilege::check_super_user;
+use super::RwPgResponse;
 use crate::binder::Binder;
 use crate::handler::drop_table::check_source;
 use crate::session::OptimizerContext;
@@ -25,7 +26,7 @@ use crate::session::OptimizerContext;
 pub async fn handle_drop_sink(
     context: OptimizerContext,
     sink_name: ObjectName,
-) -> Result<PgResponse> {
+) -> Result<RwPgResponse> {
     let session = context.session_ctx;
     let (schema_name, sink_name) = Binder::resolve_table_name(sink_name)?;