--- a/src/frontend/src/handler/flush.rs
+++ b/src/frontend/src/handler/flush.rs
@@ -15,9 +15,10 @@
 use pgwire::pg_response::{PgResponse, StatementType};
 use risingwave_common::error::Result;
 
+use super::RwPgResponse;
 use crate::session::OptimizerContext;
 
-pub(super) async fn handle_flush(context: OptimizerContext) -> Result<PgResponse> {
+pub(super) async fn handle_flush(context: OptimizerContext) -> Result<RwPgResponse> {
     let client = context.session_ctx.env().meta_client();
     // The returned epoch >= epoch for flush, but it is okay.
     let snapshot = client.flush(true).await?;