--- a/src/frontend/src/handler/mod.rs
+++ b/src/frontend/src/handler/mod.rs
@@ -22,7 +22,6 @@ use pgwire::pg_response::StatementType::{ABORT, BEGIN, COMMIT, ROLLBACK, START_T
 use pgwire::pg_response::{PgResponse, RowSetResult};
 use pgwire::pg_server::BoxedError;
 use pgwire::types::Row;
-use pin_project::pin_project;
 use risingwave_common::error::{ErrorCode, Result};
 use risingwave_sqlparser::ast::{DropStatement, ObjectType, Statement};
 
@@ -60,7 +59,6 @@ pub mod variable;
 /// The [`PgResponse`] used by Risingwave.
 pub type RwPgResponse = PgResponse<PgResponseStream>;
 
-#[pin_project(project = PgResponseStreamImpl)]
 pub enum PgResponseStream {
     LocalQuery(LocalQueryStream),
     DistributedQuery(DistributedQueryStream),
@@ -70,11 +68,11 @@ pub enum PgResponseStream {
 impl Stream for PgResponseStream {
     type Item = std::result::Result<Vec<Row>, BoxedError>;
 
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        match self.project() {
-            PgResponseStreamImpl::LocalQuery(inner) => inner.poll_next_unpin(cx),
-            PgResponseStreamImpl::DistributedQuery(inner) => inner.poll_next_unpin(cx),
-            PgResponseStreamImpl::Rows(inner) => inner.poll_next_unpin(cx),
+    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        match &mut *self {
+            PgResponseStream::LocalQuery(inner) => inner.poll_next_unpin(cx),
+            PgResponseStream::DistributedQuery(inner) => inner.poll_next_unpin(cx),
+            PgResponseStream::Rows(inner) => inner.poll_next_unpin(cx),
         }
     }
 }