@Gun9niR 2022-10-08
> Maybe we can consider add micro bench just like batch executors later. LGTM!

The difficulty is that optimizations like this involve interface changes, so it's very hard to make the bench code stable.